экран разделен на несколько отдельных компонентов
