+++ packages/client/src/pages/Game/GameStart/GameStart.tsx
@@ -1,45 +1,15 @@
 import React, { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { Flex, Radio, Typography } from 'antd'
+import { Flex, Typography } from 'antd'
 import { Button } from '../../../components/Button/Button'
 import { AppRoutes } from '../../../AppRoutes'
+import CharacterChoice from './components/CharacterChoice/CharacterChoice'
+import Rules from './components/Rules'
 import styles from './GameStart.module.css'
-
-interface Character {
-  id: string
-  name: string
-  description: string
-}
-
-const characters: Character[] = [
-  {
-    id: 'ninja',
-    name: 'Ниндзя',
-    description: 'Быстрый и ловкий боец, легко перепрыгивает препятствия.',
-  },
-  {
-    id: 'robot',
-    name: 'Робот',
-    description: 'Тяжёлый, но устойчивый к ударам и способен пробивать стены.',
-  },
-  {
-    id: 'alien',
-    name: 'Пришелец',
-    description:
-      'Может парить над землёй и использовать необычные способности.',
-  },
-]
-
-const rules = [
-  'Нажмите стрелку вверх, чтобы прыгнуть',
-  'Нажмите стрелку вниз, чтобы пригнуться',
-  'Собирайте монеты, чтобы купить жизни или способности',
-]
 
 const GameStart: React.FC = () => {
   const navigate = useNavigate()
 
-  const [selectedCharacter, setSelectedCharacter] = useState(characters[0].id)
   const [countdown, setCountdown] = useState(60)
   const [isCounting, setIsCounting] = useState(true)
 
@@ -87,52 +57,8 @@ const GameStart: React.FC = () => {
           </div>
         )}
       </Flex>
-      <div className={styles.characterSelect}>
-        <Typography.Title level={2} style={{ color: '#fff', fontSize: '28px' }}>
-          Выберите персонажа:
-        </Typography.Title>
-        <Radio.Group
-          onChange={e => setSelectedCharacter(e.target.value)}
-          value={selectedCharacter}
-          className={styles.radioGroup}>
-          {characters.map(character => (
-            <Radio key={character.id} value={character.id}>
-              {character.name}
-            </Radio>
-          ))}
-        </Radio.Group>
-        <div>
-          <Typography.Text
-            style={{
-              color: '#fff',
-              fontStyle: 'italic',
-              fontSize: '14px',
-              marginTop: '1rem',
-              display: 'block',
-            }}>
-            {
-              characters.find(character => character.id === selectedCharacter)
-                ?.description
-            }
-          </Typography.Text>
-        </div>
-      </div>
-      <Typography.Title
-        level={3}
-        style={{
-          color: '#fff',
-          fontSize: '24px',
-          textAlign: 'center',
-          marginBottom: '0',
-          marginTop: '30px',
-        }}>
-        Правила игры:
-      </Typography.Title>
-      <ul style={{ marginTop: '0' }}>
-        {rules.map((rule, index) => (
-          <li key={index}>{rule}</li>
-        ))}
-      </ul>
+      <CharacterChoice />
+      <Rules />
       <Button onClick={handleStartGame}>Старт</Button>
     </Flex>
   )
