+++ packages/client/src/pages/Game/GameStart/components/CharacterChoice/CharacterChoice.module.css
@@ -1,0 +1,7 @@
+.characterSelect {
+  text-align: center;
+}
+
+.radioGroup :global(.ant-radio-wrapper) span {
+  color: #fff;
+}
